+++ src/crops/Melon.java
@@ -7,9 +7,9 @@ public class Melon extends Crop {
 	
 	public final static String MELON_NAME = "Melon";
 	public final static String MELON_DESC = "Watermelone";
-	public final static double MELON_PRICE = 6;
-	public final static double MELON_INCOME = 300;
-	public final static double MELON_THRESHOLD = 5;	
+	public final static float MELON_PRICE = 6;
+	public final static float MELON_INCOME = 300;
+	public final static float MELON_THRESHOLD = 5;
 	
     public Melon() {
 		super(MELON_NAME, MELON_DESC, MELON_PRICE, MELON_INCOME, MELON_THRESHOLD);
